26/01/2023 eliminacion

+++ modelo/productoModelo.php
@@ -77,12 +77,21 @@ class ModeloProducto{
   }
 
   static public function mdlEliProducto($data){
-    $stmt=Conexion::conectar()->prepare("delete from producto where id_producto=$data");
-
-    if($stmt->execute()){
-      return "ok";
-    }else{
-      return "error";
+    $factura = Conexion::conectar()->prepare("select * from factura where id_producto=$data");
+    $factura->execute();
+    if($factura->fetch() > 0){
+      echo "error";
+    } else {
+      $notaEntrega = Conexion::conectar()->prepare("select * from nota_entrega id_producto=$data");
+      $notaEntrega->execute();
+      if($notaEntrega->fetch() > 0){
+        echo "error";
+      } else {
+        $stmt=Conexion::conectar()->prepare("delete from producto where id_producto=$data");
+        if($stmt->execute()){
+          return "ok";
+        }
+      }
     }
 
     $stmt->close();
